mejora de funciones

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: kcarrero <kcarrero@student.42madrid.c      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 20:42:33 by kcarrero          #+#    #+#             */
-/*   Updated: 2025/04/08 20:42:47 by kcarrero         ###   ########.fr       */
+/*   Updated: 2025/04/15 13:10:35 by kcarrero         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_strlen(const char *s)
 {
-	int	i;
-	int	count;
+	int		i;
+	size_t	count;
 
 	i = 0;
 	count = 0;
